requirements.txt
google-generativeai>=0.3.0
tavily-python>=0.1.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
